Update index.ts
If form data loading from ajax, md-select (Options example: 1, 2, 3) sync value recognize values [1, 3] only in input, but not in select dropdown list (checkboxes are not checked). If try to select options, they added to values array, so i get [1, 3, 3, 2] with dublicates, but not displaying in select input. Added methods to get unique select values and sync selected data on list open and close.